ciclo_limite_GUI/requirements.txt
opencv-python
paho-mqtt
deep-translator
Pillow